feature: allow env vars to be set when creating apps
