+++ dashboard/src/features/application/provider/applicationFormProvider.tsx
@@ -1,4 +1,4 @@
-import { useFormContext } from '../../../libs/useFormContext'
+import { useFormContext } from '/@/libs/useFormContext'
 import { createOrUpdateApplicationSchema } from '../schema/applicationSchema'
 
 export const { FormProvider: ApplicationFormProvider, useForm: useApplicationForm } = useFormContext(
