+++ dashboard/src/features/application/provider/envVarConfigFormProvider.tsx
@@ -1,0 +1,4 @@
+import { useFormContext } from '/@/libs/useFormContext'
+import { envVarSchema } from '../schema/envVarSchema'
+
+export const { FormProvider: EnvVarConfigFormProvider, useForm: useEnvVarConfigForm } = useFormContext(envVarSchema)
