+++ dashboard/src/features/repository/provider/repositoryFormProvider.tsx
@@ -1,4 +1,4 @@
-import { useFormContext } from '../../../libs/useFormContext'
+import { useFormContext } from '/@/libs/useFormContext'
 import { createOrUpdateRepositorySchema } from '../schema/repositorySchema'
 
 export const { FormProvider: RepositoryFormProvider, useForm: useRepositoryForm } =
